requirements.txt
streamlit
pillow
pytest
pytest-mock
validators